+++ v1_18_2/src/main/java/me/molybdenum/ambience_mini/v1_18_2/client/core/state/LevelState.java
@@ -9,6 +9,7 @@ import net.minecraft.client.server.IntegratedServer;
 import net.minecraft.core.BlockPos;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.animal.Animal;
 import net.minecraft.world.entity.npc.Villager;
 import net.minecraft.world.level.ClipContext;
@@ -137,6 +138,13 @@ public class LevelState extends BaseLevelState<BlockPos, Vec3, BlockState, Entit
 
 
     @Override
+    public Stream<Entity> getNearbyLiving(BlockPos center, int horizontalRadius, int verticalRadius) {
+        var nearby = cachedLevel == null ? null : getNearbyEntities(LivingEntity.class, center, horizontalRadius, verticalRadius);
+        return nearby == null ? null : nearby.stream().map(ent -> ent);
+    }
+
+
+    @Override
     public Integer getMaxSkyLightAt(BlockPos blockPos) {
         return cachedLevel == null ? null : cachedLevel.getBrightness(LightLayer.SKY, blockPos);
     }
@@ -215,6 +223,7 @@ public class LevelState extends BaseLevelState<BlockPos, Vec3, BlockState, Entit
 
     // ------------------------------------------------------------------------------------------------
     // Utilities
+    @Nullable
     private <T extends Entity> List<T> getNearbyEntities(Class<T> clazz, BlockPos center, int horizontalRadius, int verticalRadius)
     {
         var area = new AABB(