+++ v1_18_2/src/main/java/me/molybdenum/ambience_mini/v1_18_2/server/core/locations/StructureReader.java
@@ -1,6 +1,7 @@
 package me.molybdenum.ambience_mini.v1_18_2.server.core.locations;
 
 import me.molybdenum.ambience_mini.engine.server.core.locations.BaseStructureReader;
+import me.molybdenum.ambience_mini.engine.server.core.setup.BaseServerConfig;
 import me.molybdenum.ambience_mini.engine.shared.core.structures.AmStructure;
 import me.molybdenum.ambience_mini.engine.shared.utils.Pair;
 import me.molybdenum.ambience_mini.engine.shared.utils.vectors.Vector2i;
@@ -24,7 +25,8 @@ public class StructureReader extends BaseStructureReader<ServerPlayer, Level, St
     private final Registry<ConfiguredStructureFeature<?,?>> registry;
 
 
-    public StructureReader(MinecraftServer server) {
+    public StructureReader(MinecraftServer server, BaseServerConfig config) {
+        super(config);
         this.registry = server.registryAccess().registryOrThrow(Registry.CONFIGURED_STRUCTURE_FEATURE_REGISTRY);
     }